fix: refactored shell identification code
