+++ src/util/os.ts
@@ -1,5 +1,7 @@
 import WSL from 'is-wsl'
-import {homedir, platform} from 'node:os'
+import {execSync} from 'node:child_process'
+import {homedir, userInfo as osUserInfo, platform} from 'node:os'
+import path from 'node:path'
 
 /**
  * Call os.homedir() and return the result
@@ -23,4 +25,32 @@ export function getHomeDir(): string {
  */
 export function getPlatform(): 'wsl' | NodeJS.Platform {
   return WSL ? 'wsl' : platform()
+}
+
+export function getShell(): string {
+  let shellPath
+  const SHELL = process.env.SHELL ?? osUserInfo().shell?.split(path.sep)?.pop()
+  if (SHELL) {
+    shellPath = SHELL.split('/')
+  } else if (getPlatform() === 'win32') {
+    shellPath = [determineWindowsShell()]
+  } else {
+    shellPath = ['unknown']
+  }
+
+  return shellPath.at(-1) ?? 'unknown'
+}
+
+function determineWindowsShell(): string {
+  try {
+    const parentProcessName = execSync(
+      `powershell.exe -NoProfile -Command "(Get-CimInstance Win32_Process -Filter 'ProcessID = ${process.ppid}').Name"`,
+      {encoding: 'utf8'},
+    )
+    return parentProcessName.includes('powershell') || parentProcessName.includes('pwsh')
+      ? 'powershell'
+      : (process.env.COMSPEC ?? 'cmd.exe')
+  } catch {
+    return process.env.COMSPEC ?? 'cmd.exe'
+  }
 }
